Started to break up deck analysis into more files and added a Deck class
